[WIP]: Adding database functionality
Currently the database has to be empty when loading the sql_app. With
beers.db containing beers generated from IPA API, the only method that
 works is post to add a beer.

This tries to add a crud method to create beers based on the json file
instead of using the beer db. A method to allow for the db to contain
info would be the preffered fix.